添加别名git st status

+++ tools/auto_install/auto_install.py
@@ -137,6 +137,7 @@ class git_install(object):
         if(os.system('git config --global user.email 1309776181@qq.com') != 0):exit(-1)
         if(os.system('git config --global pull.rebase false') != 0):exit(-1)
         if(os.system('git config --global core.fileMode false') != 0):exit(-1)
+        if(os.system('git config --global alias.st status') != 0):exit(-1)
         if os.path.isdir(self.__nvim):
             os.system('rm -rf '+self.__nvim)
         if os.path.isdir(self.__packer):
